Updates fastcalc.pyx so that all integer numpy arrays use int64 types.
There were several places where the Cython code expected numpy
arrays with np.int or INT == long C-type.  We want all integer arrays
to be specifically int64 arrays, and this commit makes this update
within the fastcalc.pyx file.

+++ pygsti/tools/fastcalc.pyx
@@ -1255,19 +1255,19 @@ def restricted_abs_argmax(np.ndarray[double, ndim=1] ar, np.ndarray[np.int64_t, 
 #@cython.cdivision(True) # turn off divide-by-zero checking (keeping off for Python behavior of div/mods)
 @cython.boundscheck(False) # turn off bounds-checking for entire function
 @cython.wraparound(False)  # turn off negative index wrapping for entire function
-def fast_compose_cliffords(np.ndarray[INT, ndim=2] s1, np.ndarray[INT, ndim=1] p1,
-                           np.ndarray[INT, ndim=2] s2, np.ndarray[INT, ndim=1] p2):
+def fast_compose_cliffords(np.ndarray[np.int64_t, ndim=2] s1, np.ndarray[np.int64_t, ndim=1] p1,
+                           np.ndarray[np.int64_t, ndim=2] s2, np.ndarray[np.int64_t, ndim=1] p2):
     cdef INT i
     cdef INT j
     cdef INT k
     cdef INT N = s1.shape[0] // 2 # Number of qubits
 
     # Temporary space of C^T U C terms
-    cdef np.ndarray[INT, ndim=2, mode="c"] inner = np.zeros([2*N, 2*N], dtype=np.int)
+    cdef np.ndarray[np.int64_t, ndim=2, mode="c"] inner = np.zeros([2*N, 2*N], dtype=np.int)
 
     # Outputs
-    cdef np.ndarray[INT, ndim=2, mode="c"] s = np.zeros([2*N, 2*N], dtype=np.int)
-    cdef np.ndarray[INT, ndim=1, mode="c"] p = np.zeros([2*N], dtype=np.int)
+    cdef np.ndarray[np.int64_t, ndim=2, mode="c"] s = np.zeros([2*N, 2*N], dtype=np.int64)
+    cdef np.ndarray[np.int64_t, ndim=1, mode="c"] p = np.zeros([2*N], dtype=np.int64)
 
     # If C' = [[C^00, C^01], [C^10, C^11]] and U = [[0, 0], [I, 0]]
     # then C'^T U C' = [[C^10^T C00, C^10^T C^01], [C^11^T C^00, C^11^T C^01]]
